+++ errors.go
@@ -1,0 +1,10 @@
+package suzu
+
+type SuzuError struct {
+	Code    int
+	Message string
+}
+
+func (e *SuzuError) Error() string {
+	return e.Message
+}